Se arreglaron errores, se separó la clase PruebaCalculadora
